feat!: Introduce SQLite db as second level cache for workflows
- Store artifacts like workflows, workflow runs and workflow jobs in SQLite
db. This db file can be maintained across system re-deployment

+++ backend/src/main/java/de/otto/platform/gitactionboard/adapters/service/job/WorkflowsRunDetailsResponse.java
@@ -53,7 +53,7 @@ public class WorkflowsRunDetailsResponse {
 
     @NonNull
     @JsonProperty("workflow_id")
-    Integer workflowId;
+    Long workflowId;
 
     public WorkflowRun toWorkflowRun() {
       return WorkflowRun.builder()
